# contracts/openapi/billing-service-api.yml
openapi: "3.1.0"
info:
  title: Billing Microservice API
  version: "1.0.0"
  description: |
    API REST del microservicio de facturacion.

    Gestiona facturas internas del ecosistema, alimentadas por eventos
    de Guidewire BillingCenter via Apache Kafka. El campo `sourceEvent`
    permite trazar el origen del evento Kafka que creo la factura.

    ## Transiciones de estado permitidas

    | Estado actual | Estados siguientes permitidos |
    |---------------|------------------------------|
    | PENDING       | PROCESSING, CANCELLED        |
    | PROCESSING    | COMPLETED, FAILED            |
    | COMPLETED     | _(estado final)_             |
    | FAILED        | PENDING (retry)              |
    | CANCELLED     | _(estado final)_             |
  contact:
    name: Billing Microservice Team
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0

servers:
  - url: http://localhost:8083
    description: Local development server

paths:
  /api/v1/invoices:
    get:
      operationId: listInvoices
      summary: Listar facturas
      description: Retorna una lista paginada de facturas con filtros opcionales.
      tags:
        - Invoices
      parameters:
        - name: status
          in: query
          required: false
          description: Filtrar por estado de la factura
          schema:
            $ref: "#/components/schemas/InvoiceStatus"
        - name: policyId
          in: query
          required: false
          description: Filtrar por ID de la poliza
          schema:
            type: string
            format: uuid
        - name: customerId
          in: query
          required: false
          description: Filtrar por ID del cliente
          schema:
            type: string
            format: uuid
        - name: dateFrom
          in: query
          required: false
          description: Filtrar facturas desde esta fecha (inclusive)
          schema:
            type: string
            format: date
        - name: dateTo
          in: query
          required: false
          description: Filtrar facturas hasta esta fecha (inclusive)
          schema:
            type: string
            format: date
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Lista paginada de facturas
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedInvoiceList"
              example:
                content:
                  - invoiceId: "e1f2a3b4-c5d6-7890-abcd-ef1234567890"
                    policyId: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
                    customerId: "c1d2e3f4-a5b6-7890-cdef-123456789abc"
                    status: PENDING
                    totalAmount: 15000.00
                    currency: MXN
                    items:
                      - description: "Prima mensual - Seguro Auto"
                        quantity: 1
                        unitPrice: 15000.00
                        subtotal: 15000.00
                    sourceEvent: "evt-a1b2c3d4-e5f6-7890-abcd-000000000001"
                    createdAt: "2024-03-15T10:00:00Z"
                    updatedAt: "2024-03-15T10:00:00Z"
                page: 0
                size: 20
                totalElements: 1
                totalPages: 1
        "400":
          description: Solicitud invalida
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    post:
      operationId: createInvoice
      summary: Crear factura
      description: |
        Crea una nueva factura en estado PENDING.
        Normalmente invocado por el consumer de Kafka al recibir un evento
        InvoiceCreated desde BillingCenter.
      tags:
        - Invoices
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvoiceRequest"
            example:
              policyId: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
              customerId: "c1d2e3f4-a5b6-7890-cdef-123456789abc"
              items:
                - description: "Prima mensual - Seguro Auto"
                  quantity: 1
                  unitPrice: 15000.00
              sourceEvent: "evt-a1b2c3d4-e5f6-7890-abcd-000000000001"
      responses:
        "201":
          description: Factura creada exitosamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
        "400":
          description: Solicitud invalida
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: VALIDATION_ERROR
                message: "At least one item is required and totalAmount must be > 0"
                timestamp: "2024-03-15T10:00:00Z"
                path: /api/v1/invoices
        "500":
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/v1/invoices/{invoiceId}:
    get:
      operationId: getInvoiceById
      summary: Obtener factura por ID
      tags:
        - Invoices
      parameters:
        - name: invoiceId
          in: path
          required: true
          description: UUID de la factura
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Factura encontrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
        "404":
          description: Factura no encontrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateInvoice
      summary: Actualizar factura
      description: Permite actualizar el estado de una factura.
      tags:
        - Invoices
      parameters:
        - name: invoiceId
          in: path
          required: true
          description: UUID de la factura
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInvoiceRequest"
            example:
              status: PROCESSING
      responses:
        "200":
          description: Factura actualizada exitosamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
        "400":
          description: Solicitud invalida
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Factura no encontrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Transicion de estado no permitida
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: INVALID_STATE_TRANSITION
                message: "Cannot transition from COMPLETED to PROCESSING"
                timestamp: "2024-03-15T10:00:00Z"
                path: /api/v1/invoices/e1f2a3b4-c5d6-7890-abcd-ef1234567890
        "500":
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    InvoiceStatus:
      type: string
      enum:
        - PENDING
        - PROCESSING
        - COMPLETED
        - FAILED
        - CANCELLED
      description: |
        Estado del ciclo de vida de la factura del microservicio.
        - PENDING: Pendiente de procesamiento
        - PROCESSING: En proceso
        - COMPLETED: Procesamiento completado (estado final)
        - FAILED: Fallo en el procesamiento (puede reintentarse)
        - CANCELLED: Cancelada (estado final)

    Invoice:
      type: object
      required:
        - invoiceId
        - policyId
        - customerId
        - status
        - totalAmount
        - currency
        - items
        - createdAt
        - updatedAt
      properties:
        invoiceId:
          type: string
          format: uuid
          description: Identificador unico de la factura
          example: "e1f2a3b4-c5d6-7890-abcd-ef1234567890"
        policyId:
          type: string
          format: uuid
          description: UUID de la poliza asociada
          example: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
        customerId:
          type: string
          format: uuid
          description: UUID del cliente
          example: "c1d2e3f4-a5b6-7890-cdef-123456789abc"
        status:
          $ref: "#/components/schemas/InvoiceStatus"
        totalAmount:
          type: number
          format: double
          description: Monto total de la factura
          example: 15000.00
        currency:
          type: string
          description: Codigo de moneda ISO 4217
          default: "MXN"
          example: "MXN"
        items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceItem"
          description: Lineas de detalle de la factura
        sourceEvent:
          type:
            - string
            - "null"
          description: ID del evento Kafka que origino esta factura (trazabilidad)
          example: "evt-a1b2c3d4-e5f6-7890-abcd-000000000001"
        createdAt:
          type: string
          format: date-time
          description: Fecha de creacion del registro
          example: "2024-03-15T10:00:00Z"
        updatedAt:
          type: string
          format: date-time
          description: Fecha de ultima actualizacion
          example: "2024-03-15T10:00:00Z"

    InvoiceItem:
      type: object
      required:
        - description
        - quantity
        - unitPrice
        - subtotal
      properties:
        description:
          type: string
          description: Descripcion del concepto
          example: "Prima mensual - Seguro Auto"
        quantity:
          type: integer
          minimum: 1
          description: Cantidad de unidades
          example: 1
        unitPrice:
          type: number
          format: double
          minimum: 0.01
          description: Precio unitario
          example: 15000.00
        subtotal:
          type: number
          format: double
          description: Subtotal (quantity x unitPrice)
          example: 15000.00

    CreateInvoiceRequest:
      type: object
      required:
        - policyId
        - customerId
        - items
      properties:
        policyId:
          type: string
          format: uuid
          description: UUID de la poliza asociada (debe existir)
        customerId:
          type: string
          format: uuid
          description: UUID del cliente
        items:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/CreateInvoiceItemRequest"
          description: Al menos un item requerido. totalAmount se calcula automaticamente.
        sourceEvent:
          type:
            - string
            - "null"
          description: ID del evento Kafka que origina la factura

    CreateInvoiceItemRequest:
      type: object
      required:
        - description
        - quantity
        - unitPrice
      properties:
        description:
          type: string
          description: Descripcion del concepto
        quantity:
          type: integer
          minimum: 1
          description: Cantidad de unidades
        unitPrice:
          type: number
          format: double
          minimum: 0.01
          description: Precio unitario

    UpdateInvoiceRequest:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/InvoiceStatus"

    PaginatedInvoiceList:
      type: object
      required:
        - content
        - page
        - size
        - totalElements
        - totalPages
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/Invoice"
        page:
          type: integer
          example: 0
        size:
          type: integer
          example: 20
        totalElements:
          type: integer
          format: int64
          example: 8
        totalPages:
          type: integer
          example: 1

    ErrorResponse:
      type: object
      required:
        - error
        - message
        - timestamp
        - path
      properties:
        error:
          type: string
          example: "VALIDATION_ERROR"
        message:
          type: string
          example: "At least one item is required"
        timestamp:
          type: string
          format: date-time
          example: "2024-03-15T10:00:00Z"
        path:
          type: string
          example: "/api/v1/invoices"
        details:
          type: array
          items:
            type: string